backend/requirements.txt
asgiref
Django
django-allauth
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
psycopg2-binary
PyJWT
python-dotenv
pytz
sqlparse
tzdata
